Further refactoring
Expanded data type refactoring to other services

+++ services/nexus-service/src/main/java/net/thomas/portfolio/nexus/graphql/fetchers/fields/FormattedTimestampFieldDataFetcher.java
@@ -17,7 +17,7 @@ public class FormattedTimestampFieldDataFetcher extends ModelDataFetcher<String>
 
 	@Override
 	public String get(DataFetchingEnvironment environment) {
-		final Long timestamp = (Long) getEntity(environment).get(fieldName);
+		final Long timestamp = getEntity(environment).get(fieldName);
 		return dateFormatter.formatTimestamp(timestamp);
 	}
 }
